ml/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
packages = [{include = "src"}]
authors = [
    {name = "Alan Huynh",email = "hdmquan@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "loguru (>=0.7.3,<0.8.0)",
    "torch (>=2.7.0,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "numpy (>=2.2.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

